+++ src/test/java/deu/manager/executable/repository/AdminStaffJdbcRepositoryTest.java
@@ -1,15 +1,13 @@
 package deu.manager.executable.repository;
 
 import deu.manager.executable.domain.AdminStaff;
+import deu.manager.executable.repository.interfaces.AdminStaffRepository;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.data.jdbc.DataJdbcTest;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.transaction.annotation.Transactional;
